rename current computepipeline to dotspipeline
