+++ source/VulkanEngine.h
@@ -19,7 +19,7 @@
 #include "pipelines/RenderPass.h"
 #include "pipelines/objects/ObjectsPipeline.h"
 #include "pipelines/gui/GuiPipeline.h"
-#include "pipelines/compute/ComputePipeline.h"
+#include "pipelines/dots/DotsPipeline.h"
 
 #include "objects/Texture.h"
 #include "objects/Model.h"
